move rpc service into its own package
