+++ core/server/master/src/main/java/alluxio/master/AlluxioMasterProcess.java
@@ -31,10 +31,10 @@ import alluxio.master.journal.raft.RaftJournalSystem;
 import alluxio.master.journal.ufs.UfsJournalSingleMasterPrimarySelector;
 import alluxio.master.meta.DefaultMetaMaster;
 import alluxio.master.meta.MetaMaster;
-import alluxio.master.service.RpcServerSimpleService;
 import alluxio.master.service.SimpleService;
 import alluxio.master.service.jvmmonitor.JvmMonitorSimpleService;
 import alluxio.master.service.metrics.MetricsSimpleService;
+import alluxio.master.service.rpc.RpcServerSimpleService;
 import alluxio.master.service.web.WebServerSimpleService;
 import alluxio.metrics.MetricKey;
 import alluxio.metrics.MetricsSystem;
