+++ core/server/master/src/main/java/alluxio/master/AlluxioSimpleMasterProcess.java
@@ -16,8 +16,8 @@ import alluxio.conf.PropertyKey;
 import alluxio.grpc.JournalDomain;
 import alluxio.grpc.NodeState;
 import alluxio.master.journal.JournalSystem;
-import alluxio.master.service.RpcServerSimpleService;
 import alluxio.master.service.SimpleService;
+import alluxio.master.service.rpc.RpcServerSimpleService;
 import alluxio.master.service.web.WebServerSimpleService;
 import alluxio.util.network.NetworkAddressUtils;
 import alluxio.util.network.NetworkAddressUtils.ServiceType;
