+++ core/server/master/src/main/java/alluxio/master/MasterProcess.java
@@ -18,8 +18,8 @@ import alluxio.conf.AlluxioConfiguration;
 import alluxio.conf.Configuration;
 import alluxio.grpc.GrpcServerBuilder;
 import alluxio.master.journal.JournalSystem;
-import alluxio.master.service.RpcServerSimpleService;
 import alluxio.master.service.SimpleService;
+import alluxio.master.service.rpc.RpcServerSimpleService;
 import alluxio.master.service.web.WebServerSimpleService;
 import alluxio.metrics.MetricsSystem;
 import alluxio.util.CommonUtils;
